internal/vcs/git: add CommitsOptions.Before
For building historical data of search insights, I need a way to locate the commit
in a repository closest to some given point in time. For this, I am using `git.Commits`
with `Before` and `After`  to find 1 commit before and after the target time, and then
determining which is closer.

Helps #18398

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>